null safety in editText list
